Apply review remarks
